move config to assets
