--- config.js
@@ -1,1 +1,0 @@
-export const API_BASE_URL = 'https://vue-moire.skillbox.cc';
