+++ src/assets/ts/config.ts
@@ -1,0 +1,3 @@
+const API_BASE_URL = "https://vue-moire.skillbox.cc";
+
+export {API_BASE_URL};
